Factored colours and effects out of attributes.
