+++ CMakeLists.txt
@@ -64,8 +64,10 @@ set(TERMINALPP_CORE_SOURCE_FILES
     src/attribute.cpp
     src/canvas.cpp
     src/canvas_view.cpp
+    src/colour.cpp
     src/encoder.cpp
     src/glyph.cpp
+    src/effect.cpp
     src/extent.cpp
     src/point.cpp
     src/screen.cpp
@@ -86,10 +88,11 @@ set(TERMINALPP_CORE_INCLUDE_FILES
     include/terminalpp/ansi/ss3.hpp
     include/terminalpp/attribute.hpp
     include/terminalpp/behaviour.hpp
+    include/terminalpp/colour.hpp
     include/terminalpp/canvas.hpp
     include/terminalpp/canvas_view.hpp
     include/terminalpp/core.hpp
-    include/terminalpp/ansi_terminal.hpp
+    include/terminalpp/effect.hpp
     include/terminalpp/element.hpp
     include/terminalpp/encoder.hpp
     include/terminalpp/extent.hpp
@@ -148,6 +151,7 @@ set(TERMINALPP_INCLUDE_FILES
     include/terminalpp/detail/terminal_control.hpp
     include/terminalpp/detail/terminal_cursor_control.hpp
     include/terminalpp/detail/well_known_virtual_key.hpp
+    include/terminalpp/ansi_terminal.hpp
 )
 
 add_library(terminalpp
@@ -234,9 +238,6 @@ if (GTEST_FOUND)
         test/io/ansi/mouse.cpp
         test/io/string.cpp
         test/io/virtual_key.cpp
-        test/attribute_test.cpp
-        test/canvas_test.cpp
-        test/canvas_view_test.cpp
         test/ansi_terminal_init_test.cpp
         test/ansi_terminal_cursor_test.cpp
         test/ansi_terminal_read_cursor_test.cpp
@@ -244,6 +245,11 @@ if (GTEST_FOUND)
         test/ansi_terminal_read_test.cpp
         test/ansi_terminal_settings_test.cpp
         test/ansi_terminal_string_test.cpp
+        test/attribute_test.cpp
+        test/canvas_test.cpp
+        test/canvas_view_test.cpp
+        test/colour_test.cpp
+        test/effect_test.cpp
         test/element_test.cpp
         test/encoder_test.cpp
         test/expect_sequence.cpp
